refactor: 가독성을 위해 폴더 및 변수 이름 변경
