+++ src/main/java/subway/config/SubwayException.java
@@ -1,4 +1,4 @@
-package subway.util;
+package subway.config;
 
 public class SubwayException {
 }
